requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==5.0.0
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
channels==4.3.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
kombu==5.5.4
loguru==0.7.3
mysqlclient==2.2.7
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
zope.interface==8.0.1
